consider switching to doomwiki.org for wiki links/references
Hi, this is a polite, one-time request for you to please consider referencing the community-preferred "Doom Wiki" at doomwiki.org for your citations going forward.

The history: the original Doom Wiki founders decided to leave the commercial wiki host in 2011. The majority of active editors continue to develop the material over at doomwiki.org, but the commercial host refuse to delete the older wiki as a matter of policy, so it remains (and attracts a small number of edits from others).  There's a full writeup here if you are interested: https://doomwiki.org/wiki/Doom_Wiki:Departure_from_Wikia

In general the material on doomwiki is more developed (and I think that is true for this article in question, Idgames_archive).

Thanks!